requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
firebase-admin==6.2.0
requests==2.31.0
pydantic==1.10.9
python-multipart==0.0.7
cloudinary==1.32.0
twilio==8.6.0
paddleocr==2.6.1.3
paddlepaddle==2.5.0
opencv-python==4.8.1.78
Pillow==10.0.1
keras-facenet==0.3.2
tensorflow==2.13.0
mtcnn==0.1.1
